refactor: split transport tests, fix httpx client lifecycle on interrupt
- Move TestResponsesToCompletion to tests/test_codex_transport.py for
  discoverability. Add TestMessageConversion for format translation.
- Fix httpx.Client lifecycle: on interrupt, the old transport was closed
  but the rebuild reused stale _client_kwargs. Now rebuilds a fresh
  transport from current credentials.
